Created re_supersimplestocks branch
Create new branch
re_supersimplestocks
