--- src/main/java/coding/interview/supersimplestocks/stock/StockSymbol.java
@@ -1,12 +1,0 @@
-package coding.interview.supersimplestocks.stock;
-
-/**
- * Created by yarbong on 19/04/2016.
- */
-public enum StockSymbol {
-    TEA,
-    POP,
-    ALE,
-    GIN,
-    JOE
-}
