--- src/main/java/coding/interview/supersimplestocks/stock/StockType.java
@@ -1,9 +1,0 @@
-package coding.interview.supersimplestocks.stock;
-
-/**
- * Created by yarbong on 19/04/2016.
- */
-public enum StockType {
-    COMMON,
-    PREFERRED
-}
