--- src/main/java/coding/interview/supersimplestocks/trader/String.java
@@ -1,9 +1,0 @@
-package coding.interview.supersimplestocks.trader;
-
-/**
- * Created by yarbong on 19/04/2016.
- */
-public enum String {
-    BUY,
-    SELL
-}
