+++ src/main/java/supersimplestocks/stock/StockSymbol.java
@@ -1,0 +1,12 @@
+package supersimplestocks.stock;
+
+/**
+ * Created by yarbong on 19/04/2016.
+ */
+public enum StockSymbol {
+    TEA,
+    POP,
+    ALE,
+    GIN,
+    JOE
+}
