+++ src/main/java/supersimplestocks/stock/StockType.java
@@ -1,0 +1,9 @@
+package supersimplestocks.stock;
+
+/**
+ * Created by yarbong on 19/04/2016.
+ */
+public enum StockType {
+    COMMON,
+    PREFERRED
+}
