+++ src/main/java/supersimplestocks/trader/String.java
@@ -1,0 +1,9 @@
+package supersimplestocks.trader;
+
+/**
+ * Created by yarbong on 19/04/2016.
+ */
+public enum String {
+    BUY,
+    SELL
+}
